Desktop/CW2_CST1510_M01093222/Week 7-11/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
plotly>=5.0.0